Break out FontAwesome Extensions
